persist,dataflow: simplify source restore logic by assuming atomic seal
Before the recent change to use an atomic `seal()` for sealing up the
collections involved in the sink, we were working with the invariant
that the seal ts for timestamp bindings is always beyond the seal ts of
the data but that the two could be different.

Now, we use the newly introduced invariant to simplify the restore logic
which now assumes that the seal (and compaction frontier) are always the
same. We also assert that invariant.