process-orchestrator: isolate processes via uuid
This allows each environmentd process their own set of clusterds. This
is of course wrong, but its an easy way of having multiple environmentds
running on the same environment, without them clashing on the processes
that they're trying to pilot around.

One of the next steps would be to teach clusterds to listen to multiple
environmentds, but that's a task for future aljoscha.