persist: inline small writes directly in consensus State
Previously, a compare_and_append operation always required writing the
data to Blob (s3) and then only putting a pointer to it in State (stored
in Consensus/crdb). The latencies of s3 puts are both considerably
higher and with a much fatter tail than crdb latencies. So, inline small
batch parts directly where we would have put the pointer.

- The threshold for small is made a LD param, so we can tune it.
- Skip keeping pushdown stats for these parts.
- Represent them using the new LazyProto abstraction developed for part
  stats. This keeps the inline parts as bytes::Bytes (which is in turn
  just a slice of the encoded ProtoStateRollup/ProtoStateDiff), only
  decoding them when they would have been "fetched". This considerably
  lessens the cost of deserializing these in places that aren't
  interested in the updates themselves (e.g. the controllers on envd).

We expect to need considerable hardening of this before it can be rolled
out to prod. In particular, it'd be nice to somehow have hard bounds on
how big this can balloon up the size of state, though that's a bit
tricky. OTOH, this feels complete enough that we can merge it behind a
feature flag to make it easier to bash on it and find out what breaks.