compute,adapter: add peek response stash, for arbitrary-sized peek responses
Currently, results of SELECT queries (aka. peeks) are sent back from the
cluster to environmentd via the compute protocol and these results are
fully materialized (stored in memory) in environmentd before sending
them out to the client.

This has several implications:

- Sending large results "clogs up" the cluster/controller communication
- The amount of memory we want to give to environmentd limits the size
  of results we can return

In practice the above make it so we limit the size of results using a
`max_result_size` parameter, and some larger customers are chafing
against that.

Specifically, peek responses are sent inline, in the response stream
from `clusterd` to `environmentd`/Controller. As a
`ComputeResponse::PeekResponse`.

With this change, we can "stash" large peek responses and only send a
handle to them back via the response stream. Large peek responses are
stored using the new persist API for "free-standing" batches (batches
that are never intended for appending to a shard) and read out using the
also new `read_batches_consolidated`.

One current limitation is that we can only stash results when the SELECT
does _NOT_ have an ORDER BY. This is left as a follow up.

We introduces a bunch of dyncfgs for controlling when and how the result
stash is used:

- Enable/disable: for staged rollout.
- Peek stash threshold: above what result size we use the stash. This is
  deliberately separate from the existing `max_result_size`
  configuration and in practice I think we want a low-ish stash
  threshold and keep the current `max_result_size`.
- Batching while writing: we don't want to block the timely worker
  thread by writing out the whole result in one go. Instead we have
  configurable batching and batch size.
- Batching while reading: we read consolidated "chunks" out of the
  batches, when serving the result to the client. Here we also make the
  batch size configurable.

Implements https://github.com/MaterializeInc/database-issues/issues/9180

Co-authored-by: Parker Timmerman <parker.timmerman@materialize.com>